Added support for 'node paths' in the Pollo schema and displayspecification.
Now you can use things like node1/node2, so that you can specify e.g. other
attributes for node2 when it is a child of node1.

ATM, this is used in the cocoon schema to make a distinction between the
map:act element when it is a child of map:action-set and all other uses.